repana: Repeatable Analysis in R

Set of utilities to facilitate the reproduction of analysis in R. It allow to make_structure(), clean_structure(), and run and log programs in a predefined order to allow secondary files, analysis and reports be constructed in an ordered and reproducible form.

See also

Author

Maintainer: John J. Aponte john.j.aponte@gmail.com (ORCID)