Helper function to include a data.frame in the database and update the log

Helper function to include a data.frame in the database and update the log

Usage

update_table(con, table, source, tablename)

Arguments

con

DBI connection

table

the data.frame to be included in the database

source

a manual comment to identify the source of the table

tablename

if present, the data frame will be saved with this name.

Value

the result from DBI otherwise the name of the data.frame